mac files/requirements.txt
requests
python-dotenv
PyMuPDF
python-docx
json5